Factor task_report out to ds_task_report.h
